+++ filenames.gni
@@ -168,8 +168,6 @@ filenames = {
     "shell/browser/common_web_contents_delegate.h",
     "shell/browser/cookie_change_notifier.cc",
     "shell/browser/cookie_change_notifier.h",
-    "shell/browser/io_thread.cc",
-    "shell/browser/io_thread.h",
     "shell/browser/javascript_environment.cc",
     "shell/browser/javascript_environment.h",
     "shell/browser/lib/bluetooth_chooser.cc",
@@ -215,28 +213,18 @@ filenames = {
     "shell/browser/media/media_stream_devices_controller.h",
     "shell/browser/net/asar/asar_url_loader.cc",
     "shell/browser/net/asar/asar_url_loader.h",
-    "shell/browser/net/asar/url_request_asar_job.cc",
-    "shell/browser/net/asar/url_request_asar_job.h",
-    "shell/browser/net/atom_cert_verifier.cc",
-    "shell/browser/net/atom_cert_verifier.h",
-    "shell/browser/net/atom_network_delegate.cc",
-    "shell/browser/net/atom_network_delegate.h",
     "shell/browser/net/atom_url_loader_factory.cc",
     "shell/browser/net/atom_url_loader_factory.h",
-    "shell/browser/net/atom_url_request_job_factory.cc",
-    "shell/browser/net/atom_url_request_job_factory.h",
+    "shell/browser/net/cert_verifier_client.cc",
+    "shell/browser/net/cert_verifier_client.h",
     "shell/browser/net/proxying_url_loader_factory.cc",
     "shell/browser/net/proxying_url_loader_factory.h",
-    "shell/browser/net/js_asker.cc",
-    "shell/browser/net/js_asker.h",
     "shell/browser/net/network_context_service_factory.cc",
     "shell/browser/net/network_context_service_factory.h",
     "shell/browser/net/network_context_service.cc",
     "shell/browser/net/network_context_service.h",
     "shell/browser/net/node_stream_loader.cc",
     "shell/browser/net/node_stream_loader.h",
-    "shell/browser/net/require_ct_delegate.cc",
-    "shell/browser/net/require_ct_delegate.h",
     "shell/browser/net/resolve_proxy_helper.cc",
     "shell/browser/net/resolve_proxy_helper.h",
     "shell/browser/net/system_network_context_manager.cc",
