+++ shell/browser/atom_browser_client.cc
@@ -60,12 +60,12 @@
 #include "shell/browser/atom_speech_recognition_manager_delegate.h"
 #include "shell/browser/child_web_contents_tracker.h"
 #include "shell/browser/font_defaults.h"
-#include "shell/browser/io_thread.h"
 #include "shell/browser/media/media_capture_devices_dispatcher.h"
 #include "shell/browser/native_window.h"
 #include "shell/browser/net/network_context_service.h"
 #include "shell/browser/net/network_context_service_factory.h"
 #include "shell/browser/net/proxying_url_loader_factory.h"
+#include "shell/browser/net/system_network_context_manager.h"
 #include "shell/browser/notifications/notification_presenter.h"
 #include "shell/browser/notifications/platform_notification_service.h"
 #include "shell/browser/session_preferences.h"
