--- shell/browser/io_thread.cc
@@ -1,42 +1,0 @@
-// Copyright (c) 2017 GitHub, Inc.
-// Use of this source code is governed by the MIT license that can be
-// found in the LICENSE file.
-
-#include "shell/browser/io_thread.h"
-
-#include <string>
-#include <utility>
-
-#include "components/net_log/chrome_net_log.h"
-#include "content/public/browser/browser_thread.h"
-#include "content/public/browser/network_service_instance.h"
-#include "net/cert/caching_cert_verifier.h"
-#include "net/cert/cert_verifier.h"
-#include "net/cert/cert_verify_proc.h"
-#include "net/cert/multi_threaded_cert_verifier.h"
-#include "net/log/net_log_util.h"
-#include "net/proxy_resolution/proxy_resolution_service.h"
-#include "net/url_request/url_request_context.h"
-#include "services/network/network_service.h"
-#include "services/network/public/cpp/features.h"
-#include "services/network/public/cpp/network_switches.h"
-#include "services/network/public/mojom/net_log.mojom.h"
-
-using content::BrowserThread;
-
-IOThread::IOThread(
-    SystemNetworkContextManager* system_network_context_manager) {
-  BrowserThread::SetIOThreadDelegate(this);
-
-  system_network_context_manager->SetUp(
-      &network_context_request_, &network_context_params_,
-      &http_auth_static_params_, &http_auth_dynamic_params_);
-}
-
-IOThread::~IOThread() {
-  BrowserThread::SetIOThreadDelegate(nullptr);
-}
-
-void IOThread::Init() {}
-
-void IOThread::CleanUp() {}
