+++ shell/browser/net/system_network_context_manager.h
@@ -59,19 +59,6 @@ class SystemNetworkContextManager {
   // Returns default set of parameters for configuring the network service.
   network::mojom::NetworkContextParamsPtr CreateDefaultNetworkContextParams();
 
-  // Initializes |network_context_params| as needed to set up a system
-  // NetworkContext. If the network service is disabled,
-  // |network_context_request| will be for the NetworkContext used by the
-  // SystemNetworkContextManager. Otherwise, this method can still be used to
-  // help set up the IOThread's in-process URLRequestContext.
-  //
-  // Must be called before the system NetworkContext is first used.
-  void SetUp(
-      network::mojom::NetworkContextRequest* network_context_request,
-      network::mojom::NetworkContextParamsPtr* network_context_params,
-      network::mojom::HttpAuthStaticParamsPtr* http_auth_static_params,
-      network::mojom::HttpAuthDynamicParamsPtr* http_auth_dynamic_params);
-
   // Returns the System NetworkContext. May only be called after SetUp(). Does
   // any initialization of the NetworkService that may be needed when first
   // called.
@@ -104,12 +91,7 @@ class SystemNetworkContextManager {
 
   // NetworkContext using the network service, if the network service is
   // enabled. nullptr, otherwise.
-  network::mojom::NetworkContextPtr network_service_network_context_;
-
-  // This is a NetworkContext that wraps the IOThread's SystemURLRequestContext.
-  // Always initialized in SetUp, but it's only returned by Context() when the
-  // network service is disabled.
-  network::mojom::NetworkContextPtr io_thread_network_context_;
+  network::mojom::NetworkContextPtr network_context_;
 
   // URLLoaderFactory backed by the NetworkContext returned by GetContext(), so
   // consumers don't all need to create their own factory.
